update dyn.pp submodule import structure
